Framework for new system design
New system design required overhaul of previous proposed framework due to how information is recalled and stored on the database (new information we did not have previously). With a new understanding of how to use the database, the goal now is to strictly use the database for reading and writing data about users, projects, and hardware sets (done by the Project Manager)

This system still uses a Driver to mimic user input from a web page.

Currently there is no logic in ProjectManager, HardwareProcessor or ProjectProcessor and Cipher utilizes the encryption process from HW2 (which will need to be updated).